Add Cognito auth service and implement login/logout
After getting react-router working with faked authentication, it was time to start making calls out to Cognito. In order to make this work, the cognito-identity-js library and aws-sdk were added. These libraries are only referenced in the awsAuth service, as to encapsulate any Amazon specific code in one place. This should allow for greater portability, as moving to a new authentication provider will only require replacing one service file.