Add CreateUser component and call to AdminCreateUser in Cognito
In order to allow SuperAdmins to create new users, some changes were needed in the implementation of the Cognito state management. This was managed by ensuring that users logging on to the application are being correctly authenticated, and that the user groups were configured correctly in the Cognito admin interface to ensure correct permissions around the AdminCreateUser call.